+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - **Arrays**
   - Add the `merge()` function with the `MergeOption` and `NullsOption` helpers.
 - **Strings**
+  - Add the `chunk()` function : Splits a string into groups of length, separated by a separator.
   - Add the `replace()` function : Replaces all occurrences of a substring within a string, with optional Unicode (grapheme-safe) and normalization support.
   
 ## [1.0.7] - 2025-12-12